common: add enum.h for enums used in multiple headers
Currently, many enums are defined in a specific module header (such as
rcxml.h or view.h) even though they would be useful in others also.

This leads to an awkward tradeoff where either:

1. additional header dependencies are created just to pull in the enum
   (example: view.h depending on rcxml.h for view_placement_policy), or

2. to avoid additional dependencies, a generic type is used instead of
   the enum (for example, rcxml.h defining rc.window_switcher.criteria
   as uint32_t rather than lab_view_criteria to avoid a cyclical
   dependency on view.h).

We also have one case where an entire header (three-state.h) exists
just to define a single enum.

I suggest it would be better to have a common location for enums that
are used across multiple labwc headers, and propose "common/enum.h".
