--- include/common/three-state.h
@@ -1,11 +1,0 @@
-/* SPDX-License-Identifier: GPL-2.0-only */
-#ifndef LABWC_THREE_STATE_H
-#define LABWC_THREE_STATE_H
-
-enum three_state {
-	LAB_STATE_UNSPECIFIED = 0,
-	LAB_STATE_ENABLED,
-	LAB_STATE_DISABLED
-};
-
-#endif /* LABWC_THREE_STATE_H */
