+++ include/config/tablet-tool.h
@@ -3,11 +3,7 @@
 #define LABWC_TABLET_TOOL_CONFIG_H
 
 #include <stdint.h>
-
-enum motion {
-	LAB_TABLET_MOTION_ABSOLUTE = 0,
-	LAB_TABLET_MOTION_RELATIVE,
-};
+#include "common/enum.h"
 
 enum motion tablet_parse_motion(const char *name);
