+++ include/input/cursor.h
@@ -4,6 +4,7 @@
 
 #include <wlr/types/wlr_cursor.h>
 #include <wlr/util/edges.h>
+#include "common/enum.h"
 #include "ssd.h"
 
 struct view;
@@ -11,23 +12,6 @@ struct seat;
 struct server;
 struct wlr_surface;
 struct wlr_scene_node;
-enum wl_pointer_button_state;
-
-/* Cursors used internally by labwc */
-enum lab_cursors {
-	LAB_CURSOR_CLIENT = 0,
-	LAB_CURSOR_DEFAULT,
-	LAB_CURSOR_GRAB,
-	LAB_CURSOR_RESIZE_NW,
-	LAB_CURSOR_RESIZE_N,
-	LAB_CURSOR_RESIZE_NE,
-	LAB_CURSOR_RESIZE_E,
-	LAB_CURSOR_RESIZE_SE,
-	LAB_CURSOR_RESIZE_S,
-	LAB_CURSOR_RESIZE_SW,
-	LAB_CURSOR_RESIZE_W,
-	LAB_CURSOR_COUNT
-};
 
 struct cursor_context {
 	struct view *view;
